General Bugs Fix

+++ Agent/src/code/DFS.java
@@ -45,7 +45,7 @@ public class DFS {
 			return "NOSOLUTION";
 		
 		String plan = LLAPSearch.findPlan(goal);
-		double monetaryCost = goal.getState().getMoneySpent();
+		int monetaryCost = (int)goal.getState().getMoneySpent();
 		
 		System.out.println(LLAPSearch.pathToGoal(goal));
 		return plan + ";" + monetaryCost + ";" + nodesExpanded;
