+++ Agent/src/code/IDFS.java
@@ -78,7 +78,7 @@ public class IDFS {
 			return "NOSOLUTION";
 		
 		String plan = LLAPSearch.findPlan(goal);
-		double monetaryCost = goal.getState().getMoneySpent();
+		int monetaryCost = (int)goal.getState().getMoneySpent();
 		
 		System.out.println(LLAPSearch.pathToGoal(goal));
 		return plan + ";" + monetaryCost + ";" + nodesExpanded;
